backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.26.0
aiofiles==23.2.1
opencv-python==4.9.0.80
pillow==11.1.0
imagehash==4.3.1
rapidfuzz==3.6.0
python-Levenshtein==0.24.0
easyocr==1.7.1
requests==2.31.0
beautifulsoup4==4.12.3
lxml==5.1.0
google-generativeai==0.3.2
sqlalchemy==2.0.25
alembic==1.13.1
python-dotenv==1.0.0
numpy==2.2.4
scipy==1.15.2
pandas==2.2.3
pytest==7.4.4
pytest-asyncio==0.23.3
playwright==1.41.0
